+++ house-seeker-domain/src/main/java/pl/kala/houseseekerdomain/domain/model/response/dto/GetLocalityDto.java
@@ -14,7 +14,7 @@ public class GetLocalityDto {
 
     String voivodeship;
 
-    int distanceFromHome;
+    Double distanceFromHome;
 
-    int distanceFromRailStation;
+    Double distanceFromRailStation;
 }
